buffer: fix and document null-byte handling.
null-byte handling is nice but it is tricky and easy to get wrong on
our side, this is because it is not accounted in the 'used', but still
must be considered when comparing to 'capacity' if the 'flags'
SOL_BUFFER_FLAGS_NO_NULL_BYTE is not set.

The bug was visible with new sol_buffer_trim() by @cabelitos, it was
removing the last byte (the null-terminator) and valgrind was
reporting it when we use code that relies on it. The case that I saw
this was with sol-board-detect.c calling sol_util_load_file_string()
that ends calling sol_util_load_file_fd_buffer() that uses trim.

However, the 'sol_buffer_ensure()' was also wrong by comparing only
'used' to 'capacity' without taking nul byte in account. Likewise,
'sol_buffer_expand()' was not taking it in account.

These were solved and made clear in their documentation when the
null-byte is accounted or not.

A helper macro was introduced to make it easy to check for the
null-byte, the double-negation is horrible to read in the code :-)

many places were calling `nul_byte_size()` before checking
`SOL_NULL_CHECK(buf)`, that could segfault.

Many places were not accounting the null-byte, like
sol_buffer_steal_or_copy(), and sol_buffer_copy().

In others, ensure was being called with null-byte, but it is already
accounted by that function.

Many cases could avoid computing the new size + sol_buffer_ensure() by
using the new sol_buffer_expand().

sol_buffer_remove_data() couldn't work in memory-not-owned buffers,
but it has nothing to do with that. Actually it was not even doing the
resize to shrink the buffer, now this is done if memory is not owned,
otherwise just handle removing data.

Improve sol_buffer_set_slice() and sol_buffer_set_slice_at()
documentation as I lack clarity of their behavior prior to look at
their code.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com