sol-util-file: handle zero-sized reads and EAGAIN.
if we do a zero-sized read, then consider it a retry.

if we get EAGAIN, then make ret=0 so if we exceed maximum retry
attempts we do not return error. This is required to properly handle
reading from fds with unknown size, like reading from the network,
stdin or pipes.

One can later check the size of buffer and see if that was the
required amount.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com