sol-util-file: add sol_util_fill_buffer_exactly() and document.
sol_util_fill_buffer() was introduced to help cases where read-loop
was used, but some where expecting "up to" some bytes, while others
where expecting exactly a given amount.

Then make it clear that we do read up to a given amount, and a helper
that checks if the return is exactly what we wanted.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com