flow/unix-socket: use new sol_util_fill_buffer_exactly()
unix-socket expects reading exactly an amount of bytes per type. In
the case of the string, it reads the first 'size_t' then the actual
value.

The string handling was changed: it was incorrectly sending many
packets, if the partial read was happening (it shouldn't). Now it
reads exactly a given amount and send that packet.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com