io: use new sol_util_fill_buffer_exactly()
When a capacity is given, we must read exactly that amount, otherwise
we should read the whole file (usually for strings).

Also use the sol_util_load_file_fd_buffer() to fill the given buffer
instead of manually overiding its members, which could be dangerous
and wasn't respecting the flags.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com