efivarfs-storage: Remove useless check
It was using `realpath` to avoid writing to some location that
was not on `/sys/firmware/efi/efivars` - but it was fundamentally
flawed: `realpath` only check a file that already exists, and that may
not always be the case. In fact, will never be the case for first
writing.
Removed check for reading also, as again, one may read a value not yet
persisted - common case, Soletta persistence has default values for it.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>