flow/persistence: only emit packets if value changed.
A common pattern for persistence node is to connect in a loop the
setter and the persistence of the value, like to store a value to
serve via HTTP:

```
server(http-server/boolean:path=/value)
persistence(persistence/boolean:default_value=true,storage=fs,name=value)

server OUT -> IN persistence
persistence OUT -> IN server
```

If persistence keeps feeding the loop, it won't stop. This measure
should be applied to http-server and oic servers, these will come in
the next commits.