flow/http-server: only emit packets if value changed.
A common pattern for persistence node is to connect in a loop the
setter and the persistence of the value, like to store a value to
serve via HTTP:

```
server(http-server/boolean:path=/value)
persistence(persistence/boolean:default_value=true,storage=fs,name=value)

server OUT -> IN persistence
persistence OUT -> IN server
```

likewise, if there is a post of the same value, don't emit packets
since they did not change.

Take the time to document nodes behavior in their descriptions.