samples/flow: add comprehensive sample using OIC and HTTP
This sample provides a light-server that does both HTTP and OIC, as
well as saving data to disc. There are two controllers, one via HTTP
and a second via OIC, they interact using the keyboard.

It is just an initial sample and should be extended to cover more
real-world cases such as using form, gpio and persistence to do be
able to run as standalone products.