common: Rename main function for Zephyr
On microkernel builds, the application defines any number of tasks that
the system will start when execution begins. These may have any name
the user chooses.
For the nanokernel, the application must provide a main() function.

Since we require a single task to be defined for microkernels in order
for our macros to be useful, might as well have it named main() so the
macro works too for nanokernel builds.

Signed-off-by: Iván Briano <ivan.briano@intel.com>