coap: sol_socket_recvmsg() can now calculate next message's size, only.
For all sol-socket implementations so far, make it possible to just peek
what would be the next datagram's size, so that users of
sol_socket_recvmsg() can allocate the exact number of bytes to hold that
content and save memory.

This special call would not work for non-datagram-typed sockets --
FIXMEs were put in place to generate errors on the call if that is the
case, in the future. Since we're only dealing with SOL_SOCKET_UDP-typed
sockets everywhere until now, I made the missing check at the RIOT's
socket implementation too.

In the long term, we may also have a version of this function that does
not allocate at all, but just sets a string slice (or sol-buffer) to
point at an internal struct with the datagram's contents. That makes
sense at least to some small OSses, where we can work, at least on
coap's receiving side, with zero copies (we just parse the packet at the
receiving routine and we're done).

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>