coap: sol_socket_recvmsg() now can calculate next message's size, only.
For all sol-socket implementations so far, make it possible to just peek
what would be the next datagram's size, so that users of
sol_socket_recvmsg() can allocate the exact number of bytes to hold that
content and save memory.

In the long term, we may also have a version of this function that does
not allocate at all, but just sets a string slice (or sol-buffer) to
point at an internal struct with the datagram's contents. That makes
sense at least to some small OSses, where we can work, at least on
coap's receiving side, with zero copies (we just parse the packet at the
receiving routine and we're done).

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>