coap: change it to handle payload in sol_buffers.
This is the 1st iteration of trying to use less memory on CoAP buffers.
sol_coap will now, on packet creation, use a sol_buffer to handle all
the payload, and not allocate a blob with COAP_UDP_MTU bytes right away
anymore. This way, the user is able to append on that buffer to make the
user payload, so that less data hangs around at least on the sending
side. All places where a minimum amount of sending bytes could be
assured now do so.

We still have to think on how to proceed on some of the OIC/cbor paths,
since the involve users callbacks appending cbor data to a packet's
buffer (and issuing the callback *twice* just to have the cbor
phony-size-calculating calls to take place would be bad). At the
receiving side, we still allocate COAP_UDP_MTU bytes to the buffer. We
can make this better from here as well. On Linux, at least, we have the
MSG_TRUNC recvmsg() flag.

Also, the next iteration of this work will be to change de
sol_coap_packet_new() to receive a possible user-allocated buffer to use
as payload (at the receiving side this could be used as well). So, on
small OSs the direct network buffers could be passed around -- we'll
surely evolve from here.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>