build: Simplify SUBDIRS definition
The filter-out part is useless, and the line above it ends up just as
useless. The result from the $(shell find...) will be a space separated
list of paths including the /Makefile part, which are then filtered
against some of those paths without the /Makefile, so in the end,
nothign is filtered out.

Signed-off-by: Iván Briano <ivan.briano@intel.com>