build: Don't accept different *FLAGS between configuratio and built time
The CFLAGS and LDFLAGS passed by the user are used during configuration
time to check dependencies. Allowing the user to pass different flags
after configuration risks the features found at build time not matching
the ones found by the dependency resolver, breaking the build in the
best of cases.

To solve, don't use CFLAGS and LDFLAGS when setting the COMMON_*FLAGS
used by the build, preferring instead to always getting them from the
CONFIG_* values. And in order to simplify setting them, allow KConfig
to get the default of those values from variables passed by the user.

While we are at it, allow the same thing with the PREFIX value.

Signed-off-by: Iván Briano <ivan.briano@intel.com>