comms: Use tinycbor/cbor.h instead of cbor.h on includes
RIOT has a cbor.h of its own, in the same include path as every RIOT
module, so including cbor.h directly clashes.
Instead, make it possible to declare private headers on our Makefiles
that will be copied/generated to the stage directory and included from
there, with proper namespaces.

Signed-off-by: Iván Briano <ivan.briano@intel.com>