ioc: Remove init and shutdown from oic server API
We don't need to have an init and shutdown for oic server. Init can be
performed when first resource is added to the oic server and shutdown
can be performed when last resource is removed from the oic server or at
comms shutdown.

Add and remove resource internal functions were created so we can add
and remove internal resources without changing server reference counter.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>