sol-network: Add support for Bluetooth addresses
Multiple families are added because of how each family will interpret
the 'port' field:

The "raw" family will ignore that field, as it represents an address to
an endpoint that is established as soon as the physical link is
established, one example is the ATT endpoint.

The L2CAP family, the port will represent an PSM, which a range is
reserved to allocation by the Bluetooth SIG.

The RFCOMM family, port will represent an Channel, which by definition
are unallocated.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>