gpio: Match to API changes on Zephyr
The suspend/resume calls for GPIO have been removed, in favor of
specific methods the device implements that are used through other
means, not directly as we had here. These new methods are conditionally
built in if Zephyr is configured to do power management.

Eventually, the rest of the IO modules will have their APIs similarly
changed.

Signed-off-by: Iván Briano <ivan.briano@intel.com>