coap: Keep a pending context for !CON packets as well.
Now that we do not keep outgoing packets for those types after we send
them, we must at least keep, *for a while*, the pending context for the
packet, if it does not get a response. As we're documenting, a callaback
with some NULL values will be issued to the caller after that timeout
and he/she is responsible for waiting even more or not.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>