coap: Optimize context keeping for all types of packets.
We don't hold context for packets of type ACK/RESET/NONCON
anymore, *except* the users expects a reply for them, when at least a
pending reply context will be held for some time. We kept the original
contract on those sends with replies: whenever a timeout occurs or a
reply is received, it's up to the user to keep the context even more or
terminate it, via his/her callback's return value.

With these changes we're now much more memory-efficient for small
devices doing CoAP comms.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>
Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>