http-client: set error value on function return
Actually there could be two reasons to return at
this point, no memory to add to a vector or
invalid version.

It should be just a matter of returning the value
returned by add_param() function, but it's returning
a bool instead of an error value. It'll be fixed
soon as these API review changes are done anyway.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>