comms: Introduce the sol-bluetooth API
'sol-bluetooth' will provide a common API to basic Bluetooth
functionality:

 - Turning on/off the default controller present in the system;
 - Scanning for devices;
 - Establishing connections;

'sol_bt_enable()' returns a session expecting that in the future,
there could be session with different parameters for connection and
parameters.

If it makes sense, it is possible to extend the transport enum to
include a 'none' transport, that will only list the already known
devices, without doing a Bluetooth discovery. Right now, the behaviour
of 'sol_bt_start_scan()' is similar to what most mobile devices implement.

The main things still missing are configuring the local controller and
pairing new devices.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>