log: Don't call abort() directly
The abort() function in libc depends on some syscalls being available
that will not necessarily be there for all systems, so it's better to
avoid it.
Right now, we'll call abort() on Linux only, and fall back to plain
exit() for other systems. Eventually, if there's interest and proves to
be useful, we could also call specific functions from the target OS
that help with debugging.

Signed-off-by: Iván Briano <ivan.briano@intel.com>