UART: Sample refactoring.
This new sample introduces a producer and a consumer.
The consumer is lazy, so it will take some time to read the data.
The producer is very fast and a high volume of data will be produced.
From time to time the producer will cease the data production and wait
until there are no more pending data to be sent.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>