riot: Process all pending events in the same loop iteration
And try to sleep as much as possible. If we have idlers or timeouts
pending, we set the timeout to receive messages accordingly, if
there's nothing pending, call msg_receive() so the main loop blocks
until there's an event.
Previously, it was sleeping for a minimum time, then waking up even if
there's nothing to do.

Signed-off-by: Iván Briano <ivan.briano@intel.com>