oic: Make server requests callbacks async
Instead of sending the response packet just after calling server
requests callback, we have a send response method to be called
asynchronously.

Node.js bindings were updated and are working with new API, but it is
still exposing old API to users, so it is still async. JS API needs to
be fixed.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>