node.js: Copy correct soletta.node file in installation
Node.js build was failing everytime we rebuild the project. That was
happening because installing process used a find command to locate the
soletta.node file to be installed in sysroot. But after first build,
soletta.node file located in sysroot is found and we try to copy it to
itself.

Some people may not be affected by this bug because depending on locale
settings, find result order will be different

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>