oic: Improve client resource request API
Use approach similar to approach used in server API to create and send
resource requests. Instead of calling a function with one calllback to
fill data in request packet, adds a function to create a request and
another for sending it, so users can fill request data between those
calls.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>