zephyr: Don't use task_pipe from non-task contexts
As it turns out, we don't need two main loop implementations, and in
fact, one of them was entirely wrong. We can't use task_pipe from fiber
or ISR contexts.
If anyone knows what kind of drugs I was in when I initially wrote
that, I may need them again in the future... for science.

Signed-off-by: Iván Briano <ivan.briano@intel.com>