oic: Add option to make OIC server compatible with OIC 1.0 spec
Some changes were made in 1.1 spec in server payload and those changes were
already implemented in Soletta. Adding now an option to force OIC server
to create payloads as specified in OIC 1.0 spec. This is important
because clients that implements 1.0 spec can't communicate with 1.1
servers. Note that 1.1 clients support both, 1.0 and 1.1 servers.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>