cert: Fix certificate load duplication when using relative file names
sol_cert_load_from_file was trying to check if certificate was already
loaded before looking for it in disk. But, this code was incorrect
when using relative file names to load certificate.

The checker was comparing the full path of the previously loaded
certificate with the relative path of the file. Now, we are using the
full path to compare only when user is trying to load a new certificate
using the full path. And we are using the basename when user is using a
relative path.

igned-off-by: Otavio Pontes <otavio.pontes@intel.com>