persistence: add zephyr implementation
This will work for the Arduino 101 only (as soon as Iván's board name
passing to the code is in I'll make it more explicit). We're kinda
abusing the W25QXXDV flash mem usage, doing erases in whole (minimum)
sectors just to write changed entries (we have even to malloc() for
that, horrible), but there's not better way to adhere to what the flow
expects to happen for the persistence node.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>