coap: Send packet functions stated that pkt was relased, but it was not
In functions that sends CoAP packets, sol_coap_send_packet_with_reply(),
sol_coap_send_packet() and sol_coap_packet_send_notification(), pkt was
releases on success and in some cases on errors. As all calls to them
expects that the packet was released even on errors, we took this
approach and stated that clearly in documentation.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>