sol-bluetooth: Implement the agent API for BlueZ
The agent will allow to user input to be handled so pairing procedures
and the user can properly authorize and provide input.

In Bluetooth, depending on the input/output capabilities of the
device the pairing may use different procedures with different security
characteristics, the choice of the input/output capabilities
with the agent API depend on which of the agent callbacks are
implemented.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>