util-file: Add basename function that doesn't change param content
basename function from glibc may change the parameter content. For
example, basename of "/path/" will be "path" and basename of "" will be
".".

This commit adds a basename function that uses sol_str_slice in order
to prevent changing the parameter content. The only result value
that is different from basename from glibc is when path is an empty
string. The result will be an empty string too.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>