comms: Add so-netctl API
so-netctl is for implementing connection manager features.
The APIs should be implemented in Zephyr, connman and so on.
To connman, the connman D-Bus APIs are wrapped in the so-netctl
APIs. The user can use the so-netctl APIs to control connman.

To invoke the D-Bus method of connman, the async call is used,
the callback function is set to get the results of async call.

The basic so-netctl APIs are in the patch.
It provides the APIs of getting the name, ip address... of services.
The service connect/disconnect APIs and set/get offline
are implemented too.

The further sol-netctl APIs are being planned.

Signed-off-by: Wu Zheng <wu.zheng@intel.com>