coap: sol_coap_packet_add_uri_path_option fails for "/" path
The API currently fails if the URI parameter is only "/".
Although the user of the API should better check the parameter when
setting up a packet, not even calling it if this is the case,
I don`t think the API should return an error in this case either.
Maybe we can add a warning?

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>