lwm2m: bootstrap interface - Request and Finish implementation
Introduces the implementation of the Bootstrap Request and Bootstrap Finish operations
from the Bootstrap Interface.

Additionally to the implementation of the LWM2M infrastructure related to these
operations, the lwm2m-client sample was modified and is now able to use the API
in order to perform a Bootstrap Request (send a CoAP POST to /bs?ep={client_name})
with a LWM2M Bootstrap Server, as well as receive a Bootstrap Finish indication
(receive a CoAP POST at /bs).

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>