lwm2m: bootstrap interface - Sample Applications
This patch modifies the lwm2m-client application to be able to work as a Client
with or without Bootstrap Information (simulating the existence of a Factory
Bootstrap or the need of a Client- or Server-initiated Bootstrap prior).
to a regular Registration with a LWM2M Server).

Additionally to those changes in the client sample app, this patch introduces
a new sample application, lwm2m-bs-server, which works as a LWM2M Bootstrap Server
capable to deliver Bootstrap Information to certain LWM2M Clients.

These two apps should demonstrate the use of Soletta's LWM2M Boostrap Interface API,
both for Client- and Server-side.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>