lwm2m: bootstrap interface - Samples' Improvements
Improves the lwm2m-client sample, changing it to use a real/dynamic
implementation of the Security Object (including a struct to
represent its context (user_data)), instead of the old hard-coded one.

This patch is the first in the direction of solving a few problems
pointed out by valgrind during development of this functionality.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>