LWM2M: Support for Bootstrap in the Client Sample Application
This patch modifies the lwm2m-client application to be able to work as a Client
with or without Bootstrap Information (simulating the existence of a Factory
Bootstrap or the need of a Client- or Server-initiated Bootstrap prior to a
regular Registration with a LWM2M Server).

Changes to the lwm2m-client sample app include the use of real/dynamic
implementation of the Security Object (including a struct to
represent its context (user_data)), instead of the old hard-coded one.

This app should demonstrate the use of Soletta's client-side LWM2M Bootstrap
Interface API.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>