all: add typedefs to public API structs
Soletta already uses "sol_" namespace and most of the time a secondary
namespace that is module-specific. This makes the identifiers very
large. By removing the need of explicitly use "struct" the client code
has less noise.

The convention for Doxygen documentation was:

- Opaque structs appear in typedefs section of the
  documentation. @typedef doxygen marker is used.

- Public structs appear both in data structures and typedefs
  sections. The documentation is not marked so it applies to both (some
  markers were removed).

This patch includes a change in documentation of sol_http_url struct to
escape "#fragment" properly, so Doxygen stops warning. This was
preventing "make doc" to work -- due to a check done after doxygen call.

Fixes #1982.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>