CoAP: New API sol_coap_notify_by_callback
This patch introduces an API called sol_coap_notify_by_callback,
used to send notifications to CoAP Observers based on a callback
responsible for building the notification packet. This is useful
if the application using CoAP needs to know the Observer's address
before sending the actual notification (to implement LWM2M
Authorization, for instance).

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>