LWM2M: Client-side DTLS Security (Pre-Shared Key - PSK Mode)
This patch uses the sol_lwm2m_security API to implement secure
communications for an LWM2M Client.

For that end, it also adds a new parameter to the sol_lwm2m_client_new() API
to indicate if the client->coap_server created should use DTLS or not.

For the time being, only Pre-Shared Key (PSK) mode is supported.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>