LWM2M: BServer-side DTLS Security (Pre Shared Key - PSK Mode)
This patch uses the sol_lwm2m_security API to implement secure
communications for an LWM2M Bootstrap Server.

For that end, it also adds 2 new parameters to the sol_lwm2m_bootstrap_server_new()
API, to indicate which sol_lwm2m_security_mode the Bootstrap Server should
use, and to indicate which Clients' Pre-Shared Keys the Bootstrap Server
has previous knowledge of.
Use of DTLS is mandatory for the Bootstrap Server.

For the time being, only Pre-Shared Key (PSK) mode is supported.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>