Fix for sol_platform_get_machine_id
In case SOL_PLATFORM_LINUX is defined, this function expects an UUID
from the environment variable SOL_MACHINE_ID. It was only considering
UUIDs formed by 32 chars, and not 36.

This patch fixes that, leaving the length to be check by
sol_util_uuid_str_is_valid(), which already does that, and if the length
is 36, we remove the hyphens, since the returned value is expected to
contain only hex digits. Valid formats (32 chars and 36 chars + hyphens)
for UUIDs were obtained from [1].

[1] https://en.wikipedia.org/wiki/Universally_unique_identifier

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>