socket: Add Cipher Suites to sol_socket_ip_options
This patch moves the enum sol_socket_dtls_cipher from the private
file sol-socket-dtls.h to the public sol-socket.h.

It also adds two new fields to the struct sol_socket_ip_options:
 - enum sol_socket_dtls_cipher *cipher_suites;
 - uint16_t cipher_suites_len;
to indicate which DTLS Cipher Suites are supported and can be used
for communication through this socket.

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>