linux:mainloop:log: Fix multiple initializations
There is problem when sol_log_impl_init()/sol_log_impl_shutdown() is
called more than once. The global mutex _mutex is statically initialized
but soletta is destroying it on shutdown(). So, all further functions
that use this mutex (for lock() and unlock())will fail because the init
is not initializing it.

This commit remove the destroy() from the shutdown(). Because there is
no reason to init/destroy the mutex instead of let it available.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>