iio+network: ambient light and proximity sensor fbp object method sample
ambient-light-and-proximity.fbp is a sensor object that will read ambient
light and proximity based on timer interval. It have OUTPORT light, proximity
and sensor string. Sensor string is concatenate of light and proximity reading.

ambient-light-and-proximity-console.fbp print the light and proximity in
console.

ambient-light-and-proximity-mqtt.fbp send sensor string over mqtt.

Signed-off-by: Lay, Kuan Loon <kuan.loon.lay@intel.com>